Adding new pre-task feature.
This is disabled by default (for backwards compatibility), but when
enabled, allows filtering of tasks that start with the name, 'pretask'
to be hidden from the regular after_party run and status tasks.

Instead they will get run in a new 'after_party:pretask' namespace, with
the same tasks, `run` and `status` that are specific to the pretask
namespace.